Set signal of comment app in app config ready
