+++ froide/comments/apps.py
@@ -9,12 +9,15 @@ class CommentConfig(AppConfig):
     verbose_name = _("Comments")
 
     def ready(self):
+        from django_comments.signals import comment_will_be_posted
+
         from froide.account import account_canceled, account_merged
         from froide.account.export import registry
 
         account_canceled.connect(cancel_user)
         account_merged.connect(merge_user)
         registry.register(export_user_data)
+        comment_will_be_posted.connect(set_is_moderation_flag)
 
 
 def cancel_user(sender, user=None, **kwargs):
@@ -57,3 +60,14 @@ def export_user_data(user):
             ]
         ).encode("utf-8"),
     )
+
+
+def set_is_moderation_flag(comment, **kwargs):
+    from froide.foirequest.auth import can_moderate_foirequest
+
+    is_moderation = False
+    request = kwargs.get("request")
+    foi_request = comment.content_object.request
+    if can_moderate_foirequest(foi_request, request):
+        is_moderation = request.POST.get("is_moderation") == "on"
+    comment.is_moderation = is_moderation
